Add surcharge management functionality in hotel and trip forms, including validation and dynamic addition/removal of surcharges.
