+++ src/app/dashboard/hotels/[id]/edit/page.tsx
@@ -22,7 +22,7 @@ import { zodResolver } from "@hookform/resolvers/zod"
 import * as z from "zod"
 import { useRouter } from "next/navigation"
 import { useState, useEffect } from "react"
-import { Loader2 } from "lucide-react"
+import { Loader2, Plus, Trash } from "lucide-react"
 import { toast } from "sonner"
 import { apiRequest } from "@/lib/api"
 import { Hotel } from "@/types/hotels"
@@ -40,6 +40,13 @@ const hotelSchema = z.object({
     message: "Harga harus berupa angka"
   }),
   status: z.enum(["Aktif", "Non Aktif"]),
+  surcharges: z.array(z.object({
+    season: z.string().min(1, "Season harus diisi"),
+    start_date: z.string().min(1, "Tanggal mulai harus diisi"),
+    end_date: z.string().min(1, "Tanggal selesai harus diisi"),
+    surcharge_price: z.number().min(0, "Harga surcharge harus diisi"),
+    status: z.enum(["Aktif", "Non Aktif"]),
+  })).optional(),
 })
 
 export default function EditHotelPage({ params }: { params: Promise<{ id: string }> }) {
@@ -57,8 +64,23 @@ export default function EditHotelPage({ params }: { params: Promise<{ id: string
       occupancy: "Single Occupancy",
       price: "",
       status: "Aktif",
+      surcharges: [],
     }
   })
+
+  const handleAddSurcharge = () => {
+    const currentSurcharges = form.getValues("surcharges") || [];
+    form.setValue("surcharges", [
+      ...currentSurcharges,
+      {
+        season: "",
+        start_date: "",
+        end_date: "",
+        surcharge_price: 0,
+        status: "Aktif"
+      }
+    ])
+  }
 
   useEffect(() => {
     const fetchHotel = async () => {
@@ -79,6 +101,7 @@ export default function EditHotelPage({ params }: { params: Promise<{ id: string
             occupancy: response.data.occupancy,
             price: String(response.data.price),
             status: response.data.status as "Aktif" | "Non Aktif",
+            surcharges: response.data.surcharges || [],
           }
           console.log('Setting form values:', formValues)
           form.reset(formValues)
@@ -261,6 +284,154 @@ export default function EditHotelPage({ params }: { params: Promise<{ id: string
                         </FormItem>
                       )}
                     />
+                  </div>
+                </div>
+
+                {/* Surcharges */}
+                <div>
+                  <div className="flex justify-between items-center mb-6">
+                    <h2 className="text-xl font-semibold text-gray-900">Surcharges</h2>
+                    <Button
+                      type="button"
+                      variant="outline"
+                      size="sm"
+                      onClick={handleAddSurcharge}
+                    >
+                      <Plus className="w-4 h-4 mr-2" />
+                      Tambah Surcharge
+                    </Button>
+                  </div>
+                  <div className="space-y-6">
+                    {form.watch("surcharges")?.map((surcharge, sIndex) => (
+                      <div key={sIndex} className="p-4 bg-gray-50 rounded-lg space-y-6">
+                        <div className="flex justify-between items-center">
+                          <h3 className="font-medium">Surcharge {sIndex + 1}</h3>
+                          <Button
+                            type="button"
+                            variant="ghost"
+                            size="sm"
+                            onClick={() => {
+                              const currentSurcharges = form.getValues("surcharges") || [];
+                              form.setValue("surcharges", 
+                                currentSurcharges.filter((_, i) => i !== sIndex)
+                              )
+                            }}
+                          >
+                            <Trash className="w-4 h-4 text-red-500" />
+                          </Button>
+                        </div>
+                        <div className="grid grid-cols-4 gap-4">
+                          <FormField
+                            control={form.control}
+                            name={`surcharges.${sIndex}.season`}
+                            render={({ field }) => (
+                              <FormItem>
+                                <FormLabel>Season</FormLabel>
+                                <FormControl>
+                                  <Input placeholder="Contoh: Musim Panas" {...field} />
+                                </FormControl>
+                                <FormMessage />
+                              </FormItem>
+                            )}
+                          />
+                          <FormField
+                            control={form.control}
+                            name={`surcharges.${sIndex}.start_date`}
+                            render={({ field }) => (
+                              <FormItem>
+                                <FormLabel>Start Date</FormLabel>
+                                <FormControl>
+                                  <Input 
+                                    type="date"
+                                    {...field}
+                                    value={field.value ? new Date(field.value).toISOString().split('T')[0] : ''}
+                                    onChange={(e) => {
+                                      const date = e.target.value;
+                                      if (date) {
+                                        field.onChange(new Date(date).toISOString().split('T')[0]);
+                                      } else {
+                                        field.onChange('');
+                                      }
+                                    }}
+                                  />
+                                </FormControl>
+                                <FormMessage />
+                              </FormItem>
+                            )}
+                          />
+                          <FormField
+                            control={form.control}
+                            name={`surcharges.${sIndex}.end_date`}
+                            render={({ field }) => (
+                              <FormItem>
+                                <FormLabel>End Date</FormLabel>
+                                <FormControl>
+                                  <Input 
+                                    type="date"
+                                    {...field}
+                                    value={field.value ? new Date(field.value).toISOString().split('T')[0] : ''}
+                                    onChange={(e) => {
+                                      const date = e.target.value;
+                                      if (date) {
+                                        field.onChange(new Date(date).toISOString().split('T')[0]);
+                                      } else {
+                                        field.onChange('');
+                                      }
+                                    }}
+                                  />
+                                </FormControl>
+                                <FormMessage />
+                              </FormItem>
+                            )}
+                          />
+                          <FormField
+                            control={form.control}
+                            name={`surcharges.${sIndex}.surcharge_price`}
+                            render={({ field }) => (
+                              <FormItem>
+                                <FormLabel>Surcharge Price</FormLabel>
+                                <FormControl>
+                                  <Input 
+                                    type="number"
+                                    min="0"
+                                    {...field}
+                                    value={field.value || 0}
+                                    onChange={(e) => {
+                                      const value = e.target.value === '' ? 0 : parseInt(e.target.value);
+                                      field.onChange(value);
+                                    }}
+                                  />
+                                </FormControl>
+                                <FormMessage />
+                              </FormItem>
+                            )}
+                          />
+                        </div>
+                        <FormField
+                          control={form.control}
+                          name={`surcharges.${sIndex}.status`}
+                          render={({ field }) => (
+                            <FormItem>
+                              <FormLabel>Status</FormLabel>
+                              <FormControl>
+                                <Select onValueChange={field.onChange} value={field.value}>
+                                  <FormControl>
+                                    <SelectTrigger>
+                                      <SelectValue placeholder="Pilih status" />
+                                    </SelectTrigger>
+                                  </FormControl>
+                                  <SelectContent>
+                                    <SelectItem value="Aktif">Aktif</SelectItem>
+                                    <SelectItem value="Non Aktif">Non Aktif</SelectItem>
+                                  </SelectContent>
+                                </Select>
+                              </FormControl>
+                              <FormMessage />
+                            </FormItem>
+                          )}
+                        />
+                      </div>
+                    ))}
                   </div>
                 </div>
